+++ src/main/java/com/global/api/entities/enums/UpdateResourceFileType.java
@@ -1,0 +1,9 @@
+package com.global.api.entities.enums;
+
+public enum UpdateResourceFileType {
+    RESOURCE_FILE,
+    OFFLINE_RKI_KEY_FILE,
+    INSTALLATION_PAKAGE,
+    LICENSE_FILE
+
+}