+++ src/main/java/com/global/api/terminals/DeviceController.java
@@ -7,6 +7,8 @@ import com.global.api.entities.exceptions.ConfigurationException;
 import com.global.api.terminals.abstractions.*;
 import com.global.api.terminals.builders.TerminalAuthBuilder;
 import com.global.api.terminals.builders.TerminalManageBuilder;
+import com.global.api.terminals.builders.TerminalReportBuilder;
+import com.global.api.terminals.pax.responses.LocalDetailReportResponse;
 
 public abstract class DeviceController implements IDisposable {
     protected ITerminalConfiguration settings;
@@ -39,6 +41,8 @@ public abstract class DeviceController implements IDisposable {
     public abstract IDeviceInterface configureInterface() throws ConfigurationException;
     public abstract TerminalResponse processTransaction(TerminalAuthBuilder builder) throws ApiException;
     public abstract TerminalResponse manageTransaction(TerminalManageBuilder builder) throws ApiException;
+    public abstract LocalDetailReportResponse processLocalDetailReport(TerminalReportBuilder builder) throws ApiException;
+
 
     public void dispose() {
         if(_interface != null)