+++ src/test/java/com/global/api/tests/terminals/pax/PaxAdminTests.java
@@ -3,6 +3,7 @@ package com.global.api.tests.terminals.pax;
 import com.global.api.entities.enums.ConnectionModes;
 import com.global.api.entities.enums.DeviceType;
 import com.global.api.entities.enums.SafMode;
+import com.global.api.entities.enums.UpdateResourceFileType;
 import com.global.api.entities.exceptions.ApiException;
 import com.global.api.entities.exceptions.MessageException;
 import com.global.api.services.DeviceService;
@@ -14,8 +15,15 @@ import com.global.api.terminals.abstractions.ISignatureResponse;
 import com.global.api.terminals.messaging.IMessageSentInterface;
 import com.global.api.tests.terminals.hpa.RandomIdProvider;
 
+import lombok.SneakyThrows;
 import org.junit.Ignore;
 import org.junit.Test;
+
+import java.io.IOException;
+import java.net.URISyntaxException;
+import java.nio.file.Files;
+import java.nio.file.Paths;
+import java.util.Arrays;
 
 import static org.junit.Assert.*;
 import static org.junit.Assert.assertEquals;
@@ -26,11 +34,11 @@ public class PaxAdminTests {
     public PaxAdminTests() throws ApiException {
         ConnectionConfig deviceConfig = new ConnectionConfig();
         deviceConfig.setDeviceType(DeviceType.PAX_DEVICE);
-        deviceConfig.setConnectionMode(ConnectionModes.HTTP);
-        deviceConfig.setIpAddress("10.12.220.172");
+        deviceConfig.setConnectionMode(ConnectionModes.TCP_IP);
+        deviceConfig.setIpAddress("192.168.228.130");
         deviceConfig.setPort(10009);
         deviceConfig.setRequestIdProvider(new RandomIdProvider());
-        
+
         device = DeviceService.create(deviceConfig);
         assertNotNull(device);
     }
@@ -76,7 +84,8 @@ public class PaxAdminTests {
         assertEquals("OK", response.getDeviceResponseText());
     }
 
-    @Test @Ignore
+    @Test
+    @Ignore
     public void reboot() throws ApiException {
         device.setOnMessageSent(new IMessageSentInterface() {
             public void messageSent(String message) {
@@ -119,11 +128,47 @@ public class PaxAdminTests {
         assertEquals("OK", response.getDeviceResponseText());
         assertNotNull(response.getSignatureData());
     }
-    
+
     @Test
     public void enableStoreAndForwardMode() throws ApiException {
         IDeviceResponse response = device.setStoreAndForwardMode(SafMode.STAY_OFFLINE);
         assertNotNull(response);
         assertEquals("000000", response.getDeviceResponseCode());
     }
+
+    @Test
+    public void UpdateResourceFile() throws ApiException, IOException, URISyntaxException {
+        device.setOnMessageSent(new IMessageSentInterface() {
+            public void messageSent(String message) {
+                assertNotNull(message);
+                assertTrue(message.startsWith("[STX]A18[FS]1.35[FS]"));
+            }
+        });
+
+        boolean isHTTPConnectionMode = false;
+        String path = Paths.get(System.getProperty("user.dir") + "\\src\\test\\java\\com\\global\\api\\tests\\testdata\\SampleImage.zip").toString();
+
+        byte[] fileData = Files.readAllBytes(Paths.get(path));
+
+        IDeviceResponse response = device.updateResource(UpdateResourceFileType.RESOURCE_FILE, fileData, false);
+        assertEquals("OK", response.getDeviceResponseText());
+        assertNotNull(response);
+
+    }
+
+  //  @SneakyThrows
+    @Test
+    public void deleteImageFile() throws ApiException {
+        device.setOnMessageSent(new IMessageSentInterface() {
+            public void messageSent(String message) {
+                assertNotNull(message);
+                assertTrue(message.startsWith("[STX]A22[FS]1.35[FS]"));
+            }
+        });
+
+
+      IDeviceResponse response = device.deleteImage("SampleImage.png");
+      assertEquals("OK", response.getDeviceResponseText());
+      assertNotNull(response);
+  }
 }